requirements.txt
asgiref==3.6.0
certifi==2022.12.7
Django==4.2
django-autoslug==1.9.9
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-phone-field==1.8.1
sqlparse==0.4.4
